--- howhowSpecialProject/src/main/java/com/howhow/ecpayment/ECPaymentResult.java
@@ -1,5 +1,0 @@
-package com.howhow.ecpayment;
-
-public class ECPaymentResult {
-
-}
