+++ howhowSpecialProject/src/main/java/com/howhow/ecpayment/controller/ECPayResultController.java
@@ -1,0 +1,5 @@
+package com.howhow.ecpayment.controller;
+
+public class ECPayResultController {
+
+}
